+++ src/init/features/apphosting.ts
@@ -91,16 +91,8 @@ export async function doSetup(setup: Setup, config: Config, options: Options): P
       utils.logWarning(`Firebase web app not set`);
     }
 
-    const experiments = await dynamicImport("./experiments");
     const prompts = await dynamicImport("./apphosting/prompts");
-
-    let runtime = experiments.isEnabled("abiu") ? prompts.DEFAULT_RUNTIME : undefined;
-    let automaticBaseImageUpdatesDisabled = experiments.isEnabled("abiu") ? false : undefined;
-
-    if (experiments.isEnabled("abiu") && !options.nonInteractive) {
-      runtime = await prompts.promptRuntime(projectId, location);
-      automaticBaseImageUpdatesDisabled = !(await prompts.promptAutomaticBaseImageUpdates());
-    }
+    const runtime = await prompts.resolveRuntime(projectId, location, options.nonInteractive);
 
     const createBackendSpinner = ora("Creating your new backend...").start();
     const backend = await createBackend(
@@ -112,7 +104,6 @@ export async function doSetup(setup: Setup, config: Config, options: Options): P
       webApp?.id,
       /* rootDir= */ "/",
       runtime,
-      automaticBaseImageUpdatesDisabled,
     );
 
     createBackendSpinner.succeed(`Successfully created backend!\n\t${backend.name}\n`);